refactor: code cleaning for the PR#4 and better example
