+++ example/.ruby-version
@@ -1,1 +1,2 @@
-2.7.4
+2.7.5
+
