+++ example/lib/SegmentedControl.tsx
